move stf to core
