+++ server/v2/core/appmanager/types.go
